requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
mysqlclient==2.1.1
pytz==2022.2.1
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.2
urllib3==1.26.12
youtube-search-python==1.6.6
